+++ NEWS.md
@@ -1,4 +1,9 @@
-# roxytypes (development version)
+# roxytypes 0.0.0.dev
+
+* Automatically apply default formatting based on field types. For `type` 
+  fields, this will perform a default formatting 
+  (see `roxytypes:::as.character.roxy_tag_field_type`). To use another 
+  formatting style in a config, use `"{unclass(type)}"` (@dgkf #13)
 
 * Improved the stability of some custom formatting syntax, and improved
   documentation for user-provided custom formatting functions. (@dgkf #11)